+++ tests/functional/lt_test_dump_r_config.c
@@ -38,8 +38,8 @@ int lt_test_dump_r_config(void)
     LT_LOG("%s", "Initialize handle");
     LT_ASSERT(LT_OK, lt_init(&h));
 
-    LT_LOG("%s with %d", "verify_chip_and_start_secure_session()", pkey_index_0);
-    LT_ASSERT(LT_OK, verify_chip_and_start_secure_session(&h, sh0priv, sh0pub, pkey_index_0));
+    LT_LOG("%s with %d", "verify_chip_and_start_secure_session()", PAIRING_KEY_SLOT_INDEX_0);
+    LT_ASSERT(LT_OK, verify_chip_and_start_secure_session(&h, sh0priv, sh0pub, PAIRING_KEY_SLOT_INDEX_0));
 
     LT_LOG("%s", "read_whole_R_config");
     LT_ASSERT(LT_OK, read_whole_R_config(&h, &r_config));