+++ tests/functional/lt_test_engineering_sample_01_ping.c
@@ -1,6 +1,6 @@
 /**
- * @file lt_test_ping.c
- * @brief Test function which tries Ping command with max length of message with pairing slot 0
+ * @file lt_test_engineering_sample_01_ping.c
+ * @brief Test function which tries Ping command with max length of message with all 4 pairing slots
  * @author Tropic Square s.r.o.
  *
  * @license For the license see file LICENSE.txt file in the root directory of this source tree.
@@ -22,14 +22,14 @@
  *
  * @return int
  */
-int lt_test_ping(void)
+int lt_test_engineering_sample_01_ping(void)
 {
     LT_LOG(
         "  "
         "------------------------------------------------------------------------------------------------------------"
         "-");
     LT_LOG(
-        "  -------- lt_test_ping() "
+        "  -------- lt_test_engineering_sample_01_ping() "
         "-------------------------------------------------------------------------------------");
     LT_LOG(
         "  "
@@ -64,6 +64,41 @@ int lt_test_ping(void)
     LT_LOG("%s", "lt_session_abort()");
     LT_ASSERT(LT_OK, lt_session_abort(&h));
     memset(in, 0x00, PING_LEN);
+
+    // Ping with SH1
+    LT_LOG("%s with %d", "verify_chip_and_start_secure_session()", PAIRING_KEY_SLOT_INDEX_1);
+    LT_ASSERT(LT_OK, verify_chip_and_start_secure_session(&h, sh1priv, sh1pub, PAIRING_KEY_SLOT_INDEX_1));
+    LT_LOG("%s", "lt_ping() ");
+    LT_ASSERT(LT_OK, lt_ping(&h, ping_msg, in, PING_LEN));
+    LT_LOG("Asserting %d B of Ping message", PING_LEN);
+    LT_ASSERT(0, memcmp(in, ping_msg, PING_LEN));
+    LT_LOG_LINE();
+    LT_LOG("%s", "lt_session_abort()");
+    LT_ASSERT(LT_OK, lt_session_abort(&h));
+    memset(in, 0x00, PING_LEN);
+
+    // Ping with SH2
+    LT_LOG("%s with %d", "verify_chip_and_start_secure_session()", PAIRING_KEY_SLOT_INDEX_2);
+    LT_ASSERT(LT_OK, verify_chip_and_start_secure_session(&h, sh2priv, sh2pub, PAIRING_KEY_SLOT_INDEX_2));
+    LT_LOG("%s", "lt_ping() ");
+    LT_ASSERT(LT_OK, lt_ping(&h, ping_msg, in, PING_LEN));
+    LT_LOG("Asserting %d B of Ping message", PING_LEN);
+    LT_ASSERT(0, memcmp(in, ping_msg, PING_LEN));
+    LT_LOG_LINE();
+    LT_LOG("%s", "lt_session_abort()");
+    LT_ASSERT(LT_OK, lt_session_abort(&h));
+    memset(in, 0x00, PING_LEN);
+
+    // Ping with SH3
+    LT_LOG("%s with %d", "verify_chip_and_start_secure_session()", PAIRING_KEY_SLOT_INDEX_3);
+    LT_ASSERT(LT_OK, verify_chip_and_start_secure_session(&h, sh3priv, sh3pub, PAIRING_KEY_SLOT_INDEX_3));
+    LT_LOG("%s", "lt_ping() ");
+    LT_ASSERT(LT_OK, lt_ping(&h, ping_msg, in, PING_LEN));
+    LT_LOG("Asserting %d B of Ping message", PING_LEN);
+    LT_ASSERT(0, memcmp(in, ping_msg, PING_LEN));
+    LT_LOG_LINE();
+    LT_LOG("%s", "lt_session_abort()");
+    LT_ASSERT(LT_OK, lt_session_abort(&h));
 
     // Deinit handle
     LT_LOG("%s", "lt_deinit()");