+++ tests/model_based_project/main.c
@@ -5,8 +5,6 @@
  * @license For the license see file LICENSE.txt file in the root directory of this source tree.
  */
 
-#include <stdio.h>
-
 #include "libtropic_examples.h"
 #include "libtropic_functional_tests.h"
 #include "string.h"
@@ -27,26 +25,9 @@ int main(void)
     /*int*/ lt_ex_macandd();
 #endif
 
-// Test routines
-#ifdef LT_TEST_CHIP_ID
-    /*int*/ lt_test_chip_id();
+#ifdef LT_BUILD_TESTS
+#include "lt_test_registry.c.inc"
 #endif
-#ifdef LT_TEST_WRITE_PAIRING_KEYS
-    /*int*/ lt_test_write_pairing_keys();
-#endif
-#ifdef LT_TEST_WRITE_R_CONFIG
-    /*int*/ lt_test_write_r_config();
-#endif
-#ifdef LT_TEST_PING
-    /*int*/ lt_test_ping();
-#endif
-#ifdef LT_TEST_ECC_EDDSA
-    /*int*/ lt_test_ecc_eddsa();
-#endif
-#ifdef LT_TEST_R_MEM
-    /*int*/ lt_test_r_mem();
-#endif
-#ifdef LT_TEST_ERASE_R_CONFIG
-    /*int*/ lt_test_erase_r_config();
-#endif
+
+    return 0;
 }