+++ CHANGELOG.md
@@ -40,6 +40,10 @@ files `libtropic_examples.h`, `libtropic_functional_tests.h`)
 - Rename `LT_ADD_EXAMPLES` to `LT_BUILD_EXAMPLES` to be consistent with `LT_BUILD_TESTS`
   and `LT_BUILD_DOCS`
 - Always increment host's nonce in `lt_l3_decrypt_response()`
+- When `LT_BUILD_EXAMPLES` or `LT_BUILD_TESTS` are set, CMake requires to define `LT_SH0_PRIV_PATH`
+- Removed SH0 keys and pkey_index_{0,1,2,3} from `keys/keys.c` and reflected it in examples, tests
+- Added writing of pairing keys to slots 1,2,3 in `lt_test_ping` and created `lt_test_engineering_sample_01_ping`
+  from the old version of `lt_test_ping`
 
 
 ### Added
@@ -55,6 +59,9 @@ files `libtropic_examples.h`, `libtropic_functional_tests.h`)
 - added CTest handling for functional tests
 - Added `SESSION_OFF` for indicating that secure session is off and assigned it to `lt_l3_state_t.session` flag in `lt_init()`.
 - Added function `lt_l3_invalidate_host_session_data()`, which sets all data in `lt_l3_state_t` to zero and `session` to `SESSION_OFF`; it is called in `lt_session_abort` and when ciphering/deciphering of L3 command fails (wrong TAG)
+- Added CMake variable `LT_SH0_PRIV_PATH`, based on which `sh0_keys.c` is generated for examples or tests, so
+  sh0priv and sh0pub are defined
+- Added CTest handling for functional tests run against TROPIC01 model
 
 ### Fixed